backend/requirements.txt
fastapi
uvicorn
wandb
weave
python-dotenv
crewai
openai
replicate
requests 